refactor: Move mock classes to tests/helpers.js
